chore(wikipedia): add license header in example and split the option into a go file
